Added files.
Added all working files. These files:
   * compute the optimised fisher information
   * produce the graphs of the optimised fisher information
   * contain the optimsiation data itself

Note that these files have been produced over a prolonged amount of time pre-dating the git repository. The folder has been cleaned up with (in particular) old an unused files having been removed. These old files were predominantly early attempts at procedures and techniques that the files in this commit perform. They are stored in a private branch of the repository just in case they might be needed again one day, but it is not expected that anybody other than myself will ever need to access them (and even then probably not).

For practical intents and purposes, this commit will act as the (effective) start point for the repository.