Extract CSV from Maple Plots
Wrote a .mpl file to extract the data from the Maple plots and put them into .csv file suitable for use with pgfplots in LaTeX (thorugh tikz).

Included the conversion .mpl file, the generated .csv files, and the shell script to copy them to a destination folder.

Note: these csv files *should* be constructed when the plots are first created, and at some later stage plot-ti-star.mpl will be modified to do so. However becuase the plots take a long time to compute, it’s quicker (although a little annoying) for now to extract them from the already computed plot data matrices.