Modified plotDropValues
Made some fixes to plotDropValues.mpl. Additionally, created the shell script for queueing on the grid. This shell script was put in a new PBS directiory specifically for PBS scripts. In a later commit all other PBS scripts will be moved into this folder, and appropriately modified to work correctly from within it.