Fisher Information Approximation plots
Created a file to plot ti* for the Fisher Information approximation from Bean et al. (2016). This is particular to the case n=2. This required a modification to the Fisher.mpl library, which now exports FI_Approx (with only the [2] index assigned), as well as FI.

This file was run, the results were produced and are included in this commit.

Note that unlike most all other data produced to date, this file was run on my personal laptop. The symbolic computation is sufficiently fast, and the timing entirely uninteresting for the paper, that it was not worth going to the extra bother to run it on the grid.