Plot FI* data
Added code to plot FI* to the createDataTablesforPGF.mpl file. This means that FI* plot information is created when we extract the plot data for ti* plots, and exists in the same output file. This is what will happen later when all the data (both Maple Matrices, and CSV files) are created at the same time when the graphs are computed to begin with (in plot-ti-star.mpl).

Both CSV and Maple Matrix data is created and exported (to be in keeping with other files that generate plot data). Also included is a Maple worksheet to visualise the saved Matrix data (which is a modified and moved file that should not have been left in the repository after the tidyup, but nonetheless was), and a script for PBS to tun the Maple file.