Improved ti* and dropValue plot code.
Improved the plot code for ti* and dropValue plots.

The ti* plots simply had the “base=worksheetdir” option removed from FileTools[JoinPath]() because this option doesn’t work with command line Maple.

The dropValue plots were modified to accept a PLOT_RANGE argument. Additionally the plot samples points more evenly (every 0..05 distance) and relies on the adaptive algorithm to sample more poitns as needed. Finally, the already computed drop values are used to seed the plot data, since they already exist. The reason for these changes is that when n=3 the plot time was exorbitantly large, particularly when lambda was near or larger than 4. We should remain a smooth plot with these changes, but with fewer points (albiet with a smaller range for n=3).

Note that an attempt was made to initialise the remember table with the pre-computed data for dropValues used in the ti* and FI* plots. However, due to Maple making a distinction between 0.1 and 0.10 (for example), the use of a 0.05 step caused the pre-comptued values to not be detected in the remember table. So instead, we just cope with re-computing a handful of cases to a wider bound.