Fixed PBS scripts
Fixed a bug in the PBS scripts wherein a missing forward slash caused a change of directory to fail, and consequently the Maple script could not be found.

Furtheremote plotDropValues.sh needed to be brought into line with the newer scripts. I have no idea why, becuase my copy of it on a different copy of the repository has it in the correct form after the most recent commit needed to be brought into line with the newer scripts. It turns out I failed to save the changes before the most recent commit, and so they were missed (they have been added in this commit).