Updated plotdata and PBS Scripts.
Updated PBS scripts which remove the need for array jobs (makes it easier to queue and track each individual job), and changes some memory requirements.

I'm unsure why the plotdata files were modified, or why the new files for (n=3,lambda=4.0) weren't committed earlier. I'm uncertain what the difference even is. It's been over 3/4 of a year since I looked at this project. This is why I should commit as soon as I make changes.

If I were to guess, I would say that the script which produces the .csv files which are used by PGF in the paper to produce the plots in the paper just re-generated everything. Nothing substantial seems to have changed.

Regardless, these changes are being committecd to make sure they are not lost in an upcoming move of the research computing grid. If they should not have been made, this commit (or parts of it) can always be undone.