Partial Drop Detection timings for (n=5,lambda=0.5)
Drop Detection timings for (n=5,lambda=0.5). These are incomplete. I cannot find a snapshot file for this, and the timings themselves are very long (in the viscinity of a day per value of p). So I presume this computation was given up on. The timings are sufficiently bad as to make generating the actual plots simply prohibitive time-wise.

This file was added simply to complete the record (and for posterity).